New Code Revisions: Small fix
A couple small deletions that brings the total number of errors down to
178. The remaining errors I need to think more about, aside from some
javadoc issues and the previously mentioned spelling errors. I know that
some of these errors we can likely ignore and just explain why at the
demo and there are probably more of these errors that we can do this for
but not sure.