New Code Revisions: Final Fixes
Fixes javadoc issues as well as comments out code that is not actually
used in the current program. Remaining issues that occur when analyzing
the app module are spelling errors which should be ignored, an unchecked
cast that we should be able to ignore since the object being cast can't
be anything else, not using the view parameter in inherited methods
which should be ignored, two errors regarding constructors that are
never used but which is by design, and lastly a warning about the
possibility of certain graphics overlapping others if we make the text
too large but I think we can ignore this since we never change the text
resources for these. Ran a couple simple tests to make sure
functionality was preserved. Everything checked out, but since I haven't
gotten around to the map business yet I didn't check sales so that could
be checked by others if desired. I should emphasize that I ran my
analysis by checking the app module specifically since the other
analysis options brought up errors that didn't seem relevant, but I
welcome people checking these options themselves just to make sure I
didn't miss anything